Implement optional control of TBs chaining
Some AFL's QEMU patches involve proxying not only translation requests
but TB chaining requests, too. This makes QEMU run significantly faster
when using forkserver.

This patch introduces necessary callbacks to allow chaining requests
to be passed to and from an instrumentation plug-in. Particular handling
of such requests is defined by the plug-in.